Fix formula 74, 131, and 191
In the Zianis paper the unit of the volume was not properly reported. It
is not in cubic meters but in cubic decimeters according to the original
source of the formula. This commit updates the volume unit in the
docstrings for the return value and provides the original source in the
docstrings for users to verify that this unit is the correct one.