Use log, ln, exp, and pi as in Zianis paper
This commit updates the importing of the math package to rename log to
ln and log10 to log. This is done to match the names used in the Zianis
paper. Also, exp and pi are imported to shorten the formulas and match
the notation in Zianis et al. paper. Useless import math and related
comments are removed.