Make formula 204 and 205 pass tests
This commit fixes the contants c for the formulas 204 and 205 as this
makes the test cases pass. Both formulas are due to the same original
source. It seems like the same mistake was made when extracting these
two formulas. Unfortunately due missing access to the original source it
is not certain that this change adapts the formulas to their original
versions.